% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedRandIndex.R
\name{ari}
\alias{ari}
\title{Adjusted Rand Index
For comparing two clusterings or partitions
Higher scores are better. Best score is 1.
See: https://en.wikipedia.org/wiki/Rand_index}
\usage{
ari(x, y)
}
\description{
Adjusted Rand Index
For comparing two clusterings or partitions
Higher scores are better. Best score is 1.
See: https://en.wikipedia.org/wiki/Rand_index
}
