% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myimage.R
\name{color.bar.horiz}
\alias{color.bar.horiz}
\title{Adding a horizontal color bar}
\usage{
color.bar.horiz(colorVec, mn, mx = -mn, nticks = length(colorVec) - 1,
  ticks = seq(mn, mx, len = nticks), colGrids = 10, title = "",
  digits = 1)
}
\description{
Adding a horizontal color bar
}
