% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myimage.R
\name{color.bar}
\alias{color.bar}
\title{Adding a vertical color bar}
\usage{
color.bar(colorVec, mn, mx = -mn, nticks = length(colorVec) - 1,
  digits = 1, ticks = seq(mn, mx, len = nticks), title = "",
  colGrids = 10)
}
\description{
Adding a vertical color bar
}
