% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_zjk_mean.R
\name{compute_zjk_mean}
\alias{compute_zjk_mean}
\title{Compute the posterior probability that marker j in sample i is expressed}
\usage{
compute_zjk_mean(out, i, j)
}
\description{
Compute the posterior probability that marker j in sample i is expressed
}
