\name{cytof3-package}
\alias{cytof3-package}
\alias{cytof3}
\docType{package}
\title{
\packageTitle{cytof3}
}
\description{
\packageDescription{cytof3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cytof3}
\packageIndices{cytof3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cytof3}

Maintainer: \packageMaintainer{cytof3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
