% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cytof_fix_K_fit}
\alias{cytof_fix_K_fit}
\title{Cytof Model for fixed K}
\usage{
cytof_fix_K_fit(y, B, burn, warmup = 100L, thin = 1L, thin_K = 5L,
  compute_loglike_every = 1L, print_freq = 10L, ncores = 1L,
  show_timings = FALSE, prop_for_training = 0.05, shuffle_data = FALSE,
  normalize_loglike = FALSE, prior_input = NULL, truth_input = NULL,
  init_input = NULL)
}
\description{
Cytof Model for fixed K
}
