% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Z.R
\name{estimate_Z}
\alias{estimate_Z}
\title{Provides a point esitmate of Z using SALSO for feature allocation 
by David B. Dahl.}
\usage{
estimate_Z(Zs, returnIndex = FALSE)
}
\description{
Provides a point esitmate of Z using SALSO for feature allocation 
by David B. Dahl.
}
