% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Z_old.R
\name{estimate_Z_old}
\alias{estimate_Z_old}
\title{Provides a point esitmate of Z using SALSO for feature allocation 
by David B. Dahl.}
\usage{
estimate_Z_old(Zs, returnIndex = FALSE)
}
\description{
Provides a point esitmate of Z using SALSO for feature allocation 
by David B. Dahl.
}
