% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_mech_params.R
\name{gamma_params}
\alias{gamma_params}
\title{Compute parameters of gamma distribution from mean and sd
TODO: Put in different file}
\usage{
gamma_params(m, v)
}
\description{
Compute parameters of gamma distribution from mean and sd
TODO: Put in different file
}
