% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_default_fixed.R
\name{gen_default_fixed}
\alias{gen_default_fixed}
\title{Generate default fixed parameters (bool)}
\usage{
gen_default_fixed(init)
}
\arguments{
\item{init:}{list of initial values (generated by gen_default_init)}
}
\description{
Generate default fixed parameters (bool)
}
