% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_default_init.R
\name{gen_default_init}
\alias{gen_default_init}
\title{Generate default init}
\usage{
gen_default_init(prior)
}
\arguments{
\item{prior}{List of priors (from gen_default_prior(y))}
}
\description{
Generate default init
}
\examples{
init = gen_default_init(y, prior, K=10, L0=10, L1=10)
}
