% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_default_locked.R
\name{gen_default_locked}
\alias{gen_default_locked}
\title{Generate default locked (aka fixed) parameters (bool)}
\usage{
gen_default_locked(init)
}
\arguments{
\item{init:}{list of initial values (generated by gen_default_init)}
}
\description{
Generate default locked (aka fixed) parameters (bool)
}
\examples{
locked = gen_default_locked(init)
}
