% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_default_prior.R
\name{gen_default_prior}
\alias{gen_default_prior}
\title{Generate default prior}
\usage{
gen_default_prior(J, K = 10, L0 = 10, L1 = 10)
}
\description{
Generate default prior
}
\note{
TODO:\cr
- CHECK: a_sig, a_s, b_s. \cr
- Think of good priors for: alpha (a_alpha, b_alpha) \cr
}
\examples{
prior = gen_default_prior(J=32, K=10, L0=10, L1=10)
}
