% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{invgamma_params}
\alias{invgamma_params}
\title{Get the parameter values for inv-gamma distribution given mean and sd.
Parameterization: mu = b / (a-1)}
\usage{
invgamma_params(m, sig)
}
\description{
Get the parameter values for inv-gamma distribution given mean and sd.
Parameterization: mu = b / (a-1)
}
