% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_mech_params.R
\name{miss_mech_params}
\alias{miss_mech_params}
\title{Compute missing-mechanism parameters}
\usage{
miss_mech_params(y, prob)
}
\arguments{
\item{y:}{(y_lo, y_center, y_hi)}

\item{prob:}{(p_lo, p_center, p_hi)}
}
\description{
Compute missing-mechanism parameters
}
