% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Z_new.R
\name{pairwise_alloc_new}
\alias{pairwise_alloc_new}
\title{Returns a pairwise allocation matrix of a feature allocation (binary) matrix Z
using SALSO by David B. Dahl.}
\usage{
pairwise_alloc_new(Z, W, i)
}
\description{
Returns a pairwise allocation matrix of a feature allocation (binary) matrix Z
using SALSO by David B. Dahl.
}
