% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_Z_old.R
\name{pairwise_alloc_old}
\alias{pairwise_alloc_old}
\title{Returns a pairwise allocation matrix of a feature allocation (binary) matrix Z
using SALSO by David B. Dahl.}
\usage{
pairwise_alloc_old(Z)
}
\description{
Returns a pairwise allocation matrix of a feature allocation (binary) matrix Z
using SALSO by David B. Dahl.
}
