% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preimpute.R
\name{preimpute}
\alias{preimpute}
\title{Preimpute values in list of matrices y if they are missing.}
\usage{
preimpute(y, subsample_prop = 0)
}
\arguments{
\item{y}{List of matrices of same number of columns.}

\item{subsample_prop}{Proportion of original data to use (between 0 and 1). See `resample`}
}
\description{
Preimpute values in list of matrices y if they are missing.
}
