% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_mech_params.R
\name{prob_miss}
\alias{prob_miss}
\title{Compute probability of missing using missing mechanism}
\usage{
prob_miss(y, b0, b1, c0, c1)
}
\arguments{
\item{y, }{c0, b0, b1, c1}
}
\description{
Compute probability of missing using missing mechanism
}
