% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Subsample list of matrices y (if prop < 1) or upsample (if prop >= 1).}
\usage{
resample(y, prop = 1)
}
\description{
Subsample list of matrices y (if prop < 1) or upsample (if prop >= 1).
}
