% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_mech_params.R
\name{sample_from_miss_mech_prior}
\alias{sample_from_miss_mech_prior}
\title{Sample from missing mechanism prior}
\usage{
sample_from_miss_mech_prior(y, m_beta0, s2_beta0, m_beta1, s2_beta1,
  B = 10)
}
\description{
Sample from missing mechanism prior
}
