% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_init.R
\name{smart_init}
\alias{smart_init}
\title{Generate smart initial values}
\usage{
smart_init(prior, y)
}
\arguments{
\item{prior}{List of priors (from gen_default_prior(y))}

\item{y}{List of log expression levels divided by cutoffs}
}
\description{
THIS FUNCTION IS NOT READY!
}
\examples{
init = smart_init(y, prior, K=10, L0=10, L1=10)
}
