% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_gen_data_obj}
\alias{test_gen_data_obj}
\title{A test function for generating a data object from a list of matrices (y)}
\usage{
test_gen_data_obj(y)
}
\arguments{
\item{y:}{List of matrices of same dimensions.}
}
\description{
A test function for generating a data object from a list of matrices (y)
}
