% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yZ_inspect_old.R
\name{yZ_inspect_old}
\alias{yZ_inspect_old}
\title{Plot y and Z together (old version)}
\usage{
yZ_inspect_old(out, y, dat_lim, i = 0, thresh = 0.1,
  col = list(blueToRed(), greys(10))[[1]], prop_lower_panel = 0.3,
  is.postpred = FALSE, fy = function() stopifnot(TRUE),
  fZ = function() stopifnot(TRUE))
}
\arguments{
\item{fy}{function to execute after making y image}

\item{fZ}{function to execute after making Z image}
}
\description{
Plot y and Z together (old version)
}
