% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myimage.R
\name{color.bar}
\alias{color.bar}
\title{Adding a vertical color bar}
\usage{
color.bar(col, zlim = range(col), ticks = seq(zlim[1], zlim[2], len =
  length(col)), digits = 1, title = "", cex = 1)
}
\description{
Adding a vertical color bar
}
