% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_cytof_cpp}
\alias{fit_cytof_cpp}
\title{Cytof Model for fixed K
TODO: add descriptions for params}
\usage{
fit_cytof_cpp(y, B, burn, prior_ls, locked_ls, init_ls, thin = 1L,
  thin_some = 1L, compute_loglike_every = 1L, print_freq = 10L,
  ncores = 1L, joint_update_freq = 0L, use_repulsive = FALSE,
  show_timings = FALSE, normalize_loglike = FALSE, print_new_line = FALSE,
  save_gam = FALSE)
}
\arguments{
\item{y}{A list of NumericMatrix with same number of columns.}

\item{B}{(int). Number of MCMC samples (post burn-in).}

\item{burn}{(int). Burn in iterations in MCMC.}

\item{prior_ls}{Priors. Generated by `gen_default_prior`.}

\item{locked_ls}{Parameters that are not random. Generated by `gen_default_locked`.}

\item{init_ls}{Initial values. Generated by `gen_default_init`.}

\item{thin}{(int). Thinning.}

\item{thin_some}{(int). Thinning amount for some parameters.}

\item{compute_loglike_every}{(int). Frequency of computing loglike.}

\item{normalize_loglike}{(bool). Whether the log-likelihood should be normalized.}

\item{joint_update_freq(int).}{Frequency of proposing from prior (0 -> don't do it). NOT READY!}

\item{print_new_line(bool).}{Whether or not to print new line for MCMC progress}
}
\description{
Cytof Model for fixed K
TODO: add descriptions for params
}
